server/requirements.txt
Flask==2.2.2
pymongo==4.3.3
python-dotenv==0.21.0
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.4.0
pytest==7.2.0
flask_cors==3.0.10
Authlib
requests